HP-1409 Move audit log specific middleware to audit_log module
Also rename it to AuditLogMiddleware as this same middleware can do
everything needed by audit logging.
